Fix label/export pointers

+++ examples/erts/beam/beam_load.h
@@ -75,7 +75,7 @@ typedef struct {
 typedef struct {
 	Eterm function;
 	int arity;
-	uint16_t address;
+	uint32_t address;
 } ExportEntry;
 
 typedef struct {
